+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/NoConsecutiveBlankLinesRule.kt
@@ -10,7 +10,6 @@ import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 
 public class NoConsecutiveBlankLinesRule : Rule("no-consecutive-blank-lines") {
-
     override fun beforeVisitChildNodes(
         node: ASTNode,
         autoCorrect: Boolean,
@@ -20,8 +19,8 @@ public class NoConsecutiveBlankLinesRule : Rule("no-consecutive-blank-lines") {
             node.prevSibling != null
         ) {
             val text = node.getText()
-            val lfcount = text.count { it == '\n' }
-            if (lfcount < 2) {
+            val newLineCount = text.count { it == '\n' }
+            if (newLineCount < 2) {
                 return
             }
 
@@ -31,9 +30,18 @@ public class NoConsecutiveBlankLinesRule : Rule("no-consecutive-blank-lines") {
                 prevNode.treeParent.elementType == CLASS &&
                 node.treeNext.elementType == PRIMARY_CONSTRUCTOR
 
-            if (lfcount > 2 || eof || betweenClassAndPrimaryConstructor) {
+            if (newLineCount > 2 || eof || betweenClassAndPrimaryConstructor) {
                 val split = text.split("\n")
-                emit(node.startOffset + split[0].length + split[1].length + 2, "Needless blank line(s)", true)
+                val offset =
+                    node.startOffset +
+                        split[0].length +
+                        split[1].length +
+                        if (betweenClassAndPrimaryConstructor) {
+                            1
+                        } else {
+                            2
+                        }
+                emit(offset, "Needless blank line(s)", true)
                 if (autoCorrect) {
                     val newText = buildString {
                         append(split.first())